Revamp README: rebrand, add intro, update sections
Rebranded the project title to "Pokemon API" and added an introduction describing the API as an ASP.NET Core application. Updated the README to include a new "Funcionalidades" section listing the API features, and a "Como Executar" section with instructions on running the project. Removed initial test instructions, guidelines, and support information, and renamed the "Backend-end" section to "Tecnologia" with updated technology details.